WIP refactor GeodeRedisServer
